% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_palettes.R
\docType{data}
\name{lu_colours}
\alias{lu_colours}
\title{List of colours.}
\format{
An object of class \code{list} of length 3.
}
\usage{
lu_colours
}
\description{
Complete list of palettes.
Use names(lu_colours) to return all palette names.
}
\keyword{datasets}
