% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_palettes.R
\docType{data}
\name{lu_light_grey}
\alias{lu_light_grey}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
lu_light_grey
}
\description{
Define global colours
}
\keyword{datasets}
