% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_palettes.R
\name{lu_palettes}
\alias{lu_palettes}
\title{lu palettes
Generates the colour palettes}
\usage{
lu_palettes(palette_name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette_name}{Name of Palette.}

\item{n}{Number of desired colors. If number of requested colors is
beyond the scope of the palette, colours are automatically interpolated.}

\item{type}{The type of palette. One of c("discrete", "continuous").}
}
\value{
A vector of colors.
}
\description{
lu palettes
Generates the colour palettes
}
