% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUthemes-package.R
\docType{package}
\name{LUthemes-package}
\alias{LUthemes}
\alias{LUthemes-package}
\title{LUthemes: Lancaster University 'ggplot2' styling}
\description{
Provides styling ggplot2 graphics with Lancaster University colours. Unofficial theme, not endorsed by the university.
}
\author{
\strong{Maintainer}: Nicola Rennie \email{n.rennie@lancaster.ac.uk}

}
\keyword{internal}
