% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lu_palette.R
\name{set_lu_palette}
\alias{set_lu_palette}
\title{Set LU base R plotting palette}
\usage{
set_lu_palette(palette)
}
\arguments{
\item{palette}{Name of palette. See \code{names(LUthemes::lu_colours)}. Run
\code{palette("default")} to reset.}
}
\description{
Set LU base R plotting palette
}
