% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lu.R
\name{theme_lu}
\alias{theme_lu}
\title{Theme for {ggplot2} plots}
\usage{
theme_lu(
  main_family = "Roboto",
  title_family = "Roboto Slab",
  bg_colour = "white",
  text_colour = "black",
  base_size = 13
)
}
\arguments{
\item{main_family}{Name of font family to use for most of the text in the
plot. Default "Roboto".}

\item{title_family}{Name of font family to use for title of the plot.
Default "Roboto Slab".}

\item{bg_colour}{Background colour of plot. Default "white".}

\item{text_colour}{Text colour in plot. Default "black".}

\item{base_size}{Base size of text. Default 13.}
}
\description{
Theme for {ggplot2} plots
}
