% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_palettes.R
\name{view_all_palettes}
\alias{view_all_palettes}
\title{Function for printing all palettes}
\usage{
view_all_palettes()
}
\value{
A plot of all available palettes
}
\description{
Function for printing all palettes
}
