% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lu_par.R
\name{set_lu_par}
\alias{set_lu_par}
\title{Set LU base R graphical parameters}
\usage{
set_lu_par(family = "Roboto", adj = 0, mar = c(5, 3, 3, 2.5), bty = "n", ...)
}
\arguments{
\item{family}{Font used for all text elements. Default "Roboto".}

\item{adj}{Alignment of text for title. Default 0.}

\item{mar}{Margins. Default \code{c(5, 3, 3, 2.5)}.}

\item{bty}{Axis lines. Default "n".}

\item{...}{Additional arguments passed to \code{par}}
}
\description{
Set LU base R graphical parameters
}
